+++ backend/problems/src/ma1/powers.rs
@@ -1,2 +1,3 @@
 mod integer_exponents;
+mod negative_exponents;
 mod non_integer_exponents;